+++ reconstruction/src/instance/types.rs
@@ -3,12 +3,12 @@ use std::collections::HashMap;
 use crate::world::World;
 
 pub struct Types {
-    parent: HashMap<usize, usize>,
+    parent: HashMap<u32, u32>,
 }
 
 impl Types {
     pub fn new(types: spingus::domain::types::Types) -> Self {
-        let mut parent: HashMap<usize, usize> = HashMap::new();
+        let mut parent: HashMap<u32, u32> = HashMap::new();
 
         for t in types {
             let type_index = World::global().get_type_index(&t.name);
@@ -21,11 +21,11 @@ impl Types {
         Self { parent }
     }
 
-    pub fn get_parent(&self, type_index: usize) -> Option<usize> {
+    pub fn get_parent(&self, type_index: u32) -> Option<u32> {
         self.parent.get(&type_index).cloned()
     }
 
-    pub fn is_of_type(&self, type_index: usize, wished: usize) -> bool {
+    pub fn is_of_type(&self, type_index: u32, wished: u32) -> bool {
         if type_index == wished {
             return true;
         }
